requirements.txt
apturl==0.5.2
asgiref==2.3.0
asn1crypto==0.24.0
async-timeout==2.0.1
attrs==18.1.0
autobahn==18.5.1
Automat==0.6.0
beautifulsoup4==4.6.0
blinker==1.4
Brlapi==0.6.6
certifi==2018.1.18
channels==2.1.1
chardet==3.0.4
command-not-found==0.3
constantly==15.1.0
cryptography==2.1.4
cupshelpers==1.0
daphne==2.1.1
defer==1.0.6
distro-info==0.18
Django==2.0.5
django-rest-framework==0.1.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
feedparser==5.2.1
html5lib==0.999999999
httplib2==0.9.2
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
lxml==4.2.1
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.0
oauth==1.0.1
oauthlib==2.0.6
olefile==0.45.1
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.0.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
PyJWT==1.6.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyRFC3339==1.0
python-apt==1.6.0
python-debian==0.1.32
pytz==2018.4
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
simplejson==3.13.2
six==1.11.0
system-service==0.3
systemd-python==234
Twisted==18.4.0
txaio==2.10.0
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
unity-tweak-tool==0.0.7
urllib3==1.22
usb-creator==0.3.3
virtualenv==15.2.0
wadllib==1.3.2
webencodings==0.5
xkit==0.0.0
zope.interface==4.5.0
